% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_select.R
\name{gdm.variable_selection}
\alias{gdm.variable_selection}
\title{GDM variable selection}
\usage{
gdm.variable_selection(data, dst = NULL, outname = NULL, ...)
}
\arguments{
\item{data}{(string, data.frame or matrix). Input table, can be in memory (data.frame or matrix) or filepath to .CSV}

\item{dst}{(string) Filepath to write outputs to. A subdir (taking the name of outname below) will be created here. Default (NULL) will write to a temporary dir.}

\item{outname}{(string) Filename to use for writing outputs. Default (NULL) will assign name of 'gdm_variable_significance.'}

\item{...}{Other args to be passed to \code{\link{gdm.variable_significance}}}
}
\description{
Provides an automated way to select a single model using gdm.variable_significance output
}
