% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdm-variable_significance.R
\name{gdm.variable_significance}
\alias{gdm.variable_significance}
\title{GDM variable significance}
\usage{
gdm.variable_significance(data, dst, outname = "gdm_variable_significnace",
  permutations = 100, geo = TRUE, overwrite = FALSE, load_output = TRUE,
  verbose = FALSE)
}
\arguments{
\item{data}{(string, data.frame or matrix). Input table, can be in memory (data.frame or matrix) or filepath to .CSV}

\item{dst}{(string) Filepath to write outputs to. A subdir (taking the name of outname below) will be created here.}

\item{outname}{(string) Filename to use for writing outputs. Default: 'gdm_variable_selection'}

\item{permutations}{(int) Number of permutations to run for each variable. Default: 100.}

\item{geo}{(boolean) Whether or not to include geographic distance as a predictor. Default: TRUE}

\item{load_output}{(boolean) If TRUE (default) output files will be loaded into memory in a list.}

\item{verbose}{(boolean) Print out messages to console. Default: FALSE}
}
\value{
Files written to dst, or if load_output = TRUE, the output tables read in as a list.
}
\description{
Perform backward elimination to determine variable significance in a gdm
}
\examples{
gdm.variable_select(inputTable, dst_dir, analysis_1, 50, TRUE, FALSE, TRUE)

}
